refactor: 탈퇴 서비스 토큰 매니저 참조 정리
